Add example project with two competing sort functions
One is a simple insertion sort which should perform well at small input sizes. The other is a library call to std::sort. The library call may have performance optimisations similar to what Nadir provides, where e.g. it switches to insertion sort when the input size is small. If this turns out to be a problem, I may need to roll my own. Sorting functions are not too hard.