Implement my own mergesort
So std::sort was too fast. I also tried implementing a simple quicksort, but it was also too fast. Maybe I shouldn't have implemented it to sort in-place and without recursion. So then I resorted to an implementation that is known to be rather slow but with strong O(nlogn) guarantees: mergesort.

Indeed this works fantastically. Eyeballing the current result:
- For 0 and 1 the mergesort is slightly faster.
- Around 5 or 6 items, insertion sort is faster again. This is where the mergesort actually starts needing recursion and copying vectors.
- Somewhere between 100 and 500 items, mergesort takes over again due to the better time complexity.

So aside from the details around miniscule vector sizes, this seems to be a classical example of Big O notation where you have a trade-off between good time complexity or a small constant scaling.