Add problem name parameter to benchmarker
The benchmarker will separate out the problems it benchmarks options for. The user of Nadir (programmer of an application) needs to import the correct file and provide the correct variable containing measurements to Nadir's choose method.
It's impossible to store all of the measurements in one table and then have Nadir choose the correct subset based on a string, because this table needs to contain tuples of the correct parameter sets. The types and amount of these parameter sets may differ for each problem. Instead of an array you could do this in a huge tuple, with an entry for each problem, but that requires all problems to have the same amount of measurements. So this has the fewest restrictions.