Give output benchmark data a fixed name
Due to difficulties in providing the benchmark data to the choose function in a constexpr way, we'll provide it instead as static data. The static data can then still be processed in compile time, but it doesn't need to be provided to the function through trickery with template arguments or anything, and current compilers will support it properly.